src/test/java/assignment/algorithm/PathFinderTest.java
package assignment.algorithm;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.hasItem;
import static org.junit.jupiter.api.Assertions.assertEquals;

import assignment.algorithm.datastructure.PathState;
import assignment.algorithm.datastructure.Route;
import assignment.algorithm.datastructure.Star;
import assignment.algorithm.util.PathFinderHelper;
import com.google.common.graph.ValueGraph;
import com.google.common.graph.ValueGraphBuilder;
import java.util.LinkedList;
import java.util.List;
import java.util.stream.Collectors;
import java.util.stream.Stream;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;


class PathFinderTest {

  final Star a = Star.builder().id("A").name("Solar System").build();
  final Star b = Star.builder().id("B").name("Alpha Centauri").build();
  final Star c = Star.builder().id("C").name("Sirius").build();
  final Star d = Star.builder().id("D").name("Betelgeuse").build();
  final Star e = Star.builder().id("E").name("Vega").build();

  final PathFinder dfs = new PathFinder(createGraph());

  @Test
  @DisplayName("The distance of route Solar System -> Alpha Centauri -> Sirius")
  void testExercise1() {
    //when
    Route expected = Route.builder().path(new LinkedList<>(List.of(a, b, c))).totalTravelTime(9)
        .build();
    //Then
    List<Route> routes = dfs.findPath(a, c);

    //Verify
    assertEquals(PathState.VALID.getMessage(),
        PathFinderHelper.getInstance().checkPathState(routes, expected));
    assertThat(routes, hasItem(expected));
  }

  @Test
  @DisplayName("The distance of route Solar System -> Betelgeuse")
  void testExercise2() {
    //When
    Route expected = Route.builder().path(new LinkedList<>(List.of(a, d))).totalTravelTime(5)
        .build();
    //Then
    List<Route> routes = dfs.findPath(a, d);

    //Verify
    assertEquals(PathState.VALID.getMessage(),
        PathFinderHelper.getInstance().checkPathState(routes, expected));

    assertThat(routes, hasItem(expected));

  }

  @Test
  @DisplayName("The distance of route Solar System -> Betelgeuse -> Sirius")
  void testExercise3() {
    //When
    Route expected = Route.builder().path(new LinkedList<>(List.of(a, d, c))).totalTravelTime(13)
        .build();
    //Then
    List<Route> routes = dfs.findPath(a, c);

    //Verify
    assertEquals(PathState.VALID.getMessage(),
        PathFinderHelper.getInstance().checkPathState(routes, expected));

    assertThat(routes, hasItem(expected));

  }

  @Test
  @DisplayName("The distance of route Solar System -> Vega -> Alpha Centauri -> Sirius -> Betelgeuse")
  void testExercise4() {
    //When
    Route expected = Route.builder().path(new LinkedList<>(List.of(a, e, b, c, d)))
        .totalTravelTime(22).build();
    //Then
    List<Route> routes = dfs.findPath(a, d);

    //Verify
    assertEquals(PathState.VALID.getMessage(),
        PathFinderHelper.getInstance().checkPathState(routes, expected));

    assertThat(routes, hasItem(expected));

  }

  @Test
  @DisplayName("The distance of route Solar System -> Vega -> Betelgeuse")
  void testExercise5() {
    //When
    Route expected = Route.builder().path(new LinkedList<>(List.of(a, e, d))).totalTravelTime(13)
        .build();
    //Then
    List<Route> routes = dfs.findPath(a, d);

    //Verify
    assertEquals(PathState.INVALID.getMessage(),
        PathFinderHelper.getInstance().checkPathState(routes, expected));
  }

  @Test
  @DisplayName("Determine all routes starting at Sirius and ending at Sirius with a maximum of 3 stops")
  void testExercise6() {
    //When
    Route expectedRouteTwoStops = Route.builder().path(new LinkedList<>(List.of(c, d, c)))
        .totalTravelTime(16).build();
    Route expectedRouteThreeStops = Route.builder().path(new LinkedList<>(List.of(c, e, b, c)))
        .totalTravelTime(9).build();
    List<Route> expectedRoutes = List.of(expectedRouteThreeStops, expectedRouteTwoStops);

    //Then
    List<Route> routes = dfs.findPath(c, c);

    //Verify
    List<Route> routesByNumberOfStops = PathFinderHelper.getInstance()
        .getRoutesByMaxNumberOfStops(routes, 3);

    assertThat(routesByNumberOfStops, equalTo(expectedRoutes));
  }

  @Test
  @DisplayName("Determine the number of routes starting at the solar system and ending at Sirius with exactly 3 stops inbetween")
  void testExercise7() {
    //When
    Route expectedRouteThreeStops = Route.builder().path(new LinkedList<>(List.of(a, d, e, b, c)))
        .totalTravelTime(18).build();
    List<Route> expectedRoutes = List.of(expectedRouteThreeStops);

    //Then
    List<Route> routes = dfs.findPath(a, c);

    //Verify
    List<Route> routesByNumberOfStops = PathFinderHelper.getInstance()
        .getRoutesByNumberOfIntermediateStops(routes, 3);
    //TODO: I should change the implementation to support detour again when the destination is reached.
    // In this testcase: two paths are reaching to Sirius and then continue searching again to reach Sirius
    // I believe this is not addressed in graph DFS algorithm which I have implemented. So the solution seemed
    // incomplete for this testcase.
    assertThat(routesByNumberOfStops, equalTo(expectedRoutes));
  }

  @Test
  @DisplayName("Determine the duration of the shortest routes (in traveltime) between solar system and Sirius")
  void testExercise8() {
    //Execute
    Route shortestPath = dfs.findShortestPath(a, c);
    //Verify
    assertEquals(9, shortestPath.getTotalTravelTime());
  }

  @Test
  @DisplayName("Determine the duration of the shortest routes (in traveltime) starting at Alpha Centauri and ending at Alpha Centauri")
  void testExercise9() {
    //Execute
    Route path = dfs.findShortestPathInCycle(b, b);
    //verify
    assertEquals(9, path.getTotalTravelTime());
  }

  @Test
  @DisplayName("Determine all different routes starting at Sirius and ending at Sirius with an over traveltime less than 30")
  void testExercise10() {
    //When
    Route route1 = Route.builder().path(new LinkedList<>(List.of(c, d, c)))
        .totalTravelTime(16).build();
    Route route2 = Route.builder().path(new LinkedList<>(List.of(c, d, e, b, c)))
        .totalTravelTime(21).build();
    Route route3 = Route.builder().path(new LinkedList<>(List.of(c, e, b, c)))
        .totalTravelTime(9).build();
    List<Route> expectedRoutes = Stream.of(route1, route2, route3).sorted().collect(
        Collectors.toUnmodifiableList());

    //Then
    List<Route> routes = dfs.findPath(c, c);
    List<Route> routesByNumberOfStops = PathFinderHelper.getInstance()
        .getRoutesByTravelTimeLessThanLimit(routes, 30);
    //TODO: I should change the implementation to support detour again when the destination is reached.
    // In this testcase: two paths are reaching to Sirius and then continue searching again to reach Sirius
    // I believe this is not addressed in graph DFS algorithm which I have implemented. So the solution seemed
    // incomplete for this testcase.

    //Verify
    assertThat(routesByNumberOfStops, equalTo(expectedRoutes));
  }

  private ValueGraph<Star, Integer> createGraph() {
    return ValueGraphBuilder.directed()
        .<Star, Integer>immutable()
        .putEdgeValue(a, b, 5)
        .putEdgeValue(b, c, 4)
        .putEdgeValue(c, d, 8)
        .putEdgeValue(d, c, 8)
        .putEdgeValue(d, e, 6)
        .putEdgeValue(a, d, 5)
        .putEdgeValue(c, e, 2)
        .putEdgeValue(e, b, 3)
        .putEdgeValue(a, e, 7).build();
  }
}